Change the path to the system root
The physical root of the installed system will stay mounted at
/mnt/sysimage. However, the system root of the installed system
will be always mounted at /mnt/sysroot now.

These paths are defined by the Anaconda configuration and cannot
be changed during the installation. The function setSysroot only
remounts the given system root to /mnt/sysroot.

This will simplify comunication with the DBus modules, because we
don't have to inform them about the current path to the system root
anymore. It is safe to use getSysroot and setSysroot in the modules
now.