Add ImagePath property to the Live OS handler
Live OS payload handler have to find out path to the installation image.
This path is set by the "--method=livecd:<path>" from the liveinst
script. The problem is that this command is not part of the pykickstart
and it is changed in the anaconda to the harddrive installation.

Leave this logic to the main thread to make the correct decision and set
the image path properly.