Move graphical login detection to a DBus Task (#1722950)
Previously Anaconda would detect if a package provides
service(graphical-login) in the payload module and then
set the default system target accordingly - provided
it has not already been set from kickstart.

So let's move the graphical login detection code into the
DBus task that sets the system default target during the installation,
as the task has all the information it needs to achieve this.

Also tweak how we work with the default target value is-set
detection.

Resolves: rhbz#1722950