Fix default value for pwpolicy emptyok (#1664704)
The pwpolicy command supported in the %anaconda
section of kickstart uses KSOptionParser provided
by the Pykickstart module.

In RHEL7 Pykickstart still uses OptionParser for option
parsing, while on RHEL8 & in Fedora Pykickstart now uses
Argparse. This has apprently caused a subtle regression
in the default value of the emptyok option of the
pwpolicy command. On RHEL7 the default value was True
which on RHEL8 it changed to False.

So explicitely set the default value of the emptyok option
to be True, to fix this issue.

Resolves: rhbz#1664704