liveinst: Add support for plain squashfs root filesystem
livemedia-creator-31.9-1 can now create a live iso where the
LiveOS/squashfs.img is the root filesystem instead of wrapping an ext4
filesystem inside of it.

dracut's dmsquash-live module already knows how to handle this and mount
an overlayfs on top of it.

After booting the base filesystem will be mounted at /run/rootfsbase,
this examines that mountpoint and sets the LIVE_BLOCK device, the device
to use as the source for the live install, to the mounted loop device.