Increase network timeout constant
The original value (45) is not recommended for the use in as timeout for
the requests connection because it's multiple of 3.

See requests documentation:

https://3.python-requests.org/user/advanced/#timeouts

Increase to 46 to make the situation better.

This unfortunately change also three timeouts in the network.py which
is a side effect but I'm not convinced we want to create a new constant
for just 1 second difference.