Split flatpak setup to two initialization methods (#1734970)
The reason for the split is that this method will be used multiple times
to change installed repository. We need this because even the initialization
will create an installed repository and we have to make the initialization
to be able to read the remote repository. However, if the system
repository is created before the partitioning is done then the
repository is re-mounted and flatpak installation will fail.

Now there is:
initialize_with_system_path
initialize_with_path (set where you want to install)

Related: rhbz#1734970