Add flatpak cleanup method (#1734970)
This method will remote the repository and invalidate flatpak objects.

The main reason for this is that even the initialization call will
create a repository and we have to call it to be able to read remote
repository. So this will enable us to cleanup the temp repository we
will create.

Related: rhbz#1734970