Add support for DBus containers
Publishable is an abstract class for Python objects that can be
published on DBus.

A DBus container should be used to dynamically publish Publishable
objects within the same namespace. It generates a unique DBus path
for each object. It is able to resolve a DBus path into an object
and an object into a DBus path.