Add flatpak error reporting implementation (#1734970)
Flatpak run (install) call raises an exception so wrap this and
propagate it to the rpmostreepayload. Also log error callback output.

There is a question how we should react on the failed installation.
Should we kill the installation or ask user and continue? They can
install the flatpaks on their own later. Right now using the strict kill
installation.

Related: rhbz#1734970