Move progress messaging from flatpak to rpmostree (#1734970)
The reason is that rpmostree is controlling the flatpak so it has
better knowledge of what will go next. The only part of the reporting
which is still part of the flatpak is reporting about what ref is
installing. This will cluster the messages source a bit but it's much
easier for implementation and it will change because of modules soon
so there is no need to have it perfect now.

Related: rhbz#1734970