Merge rpmostreepayload installation under one task (#1734970)
There were two methods as two installation tasks:

- install
- prepare_mount_targets

Change this to just one install task with three phases:

- install
- prepare_mount_targets
- flatpak install (if available)

The main reason is that flatpak have to be run after the
prepare_mount_targets method otherwise the /var will be over-mounted.

This will break to multiple tasks on the DBus solution but for now this
is the simplest solution to handle the flatpaks correctly.

Related: rhbz#1734970