Fix noverifyssl when downloading .treeinfo file (#1723811)
Fix ssl error when downloading .treeinfo. The ssl error happens because
the ssl check is not disabled even when inst.noverifyssl is used.

When 'cacert' is None then the ssl_verify variable results to None (the
bug). However, the real problem is that None for the session.get means True...

This was a problem mainly for unified source because when the treeinfo
file is not found, the place where the .treeinfo should be taken is used
as a base repository. In other words when .treeinfo is not downloaded
no redirect is used for base repo and no additional repositories are
added from the treeinfo file.

Resolves: rhbz#1723811
Resolves: rhbz#1726156