Add base payload source module
Add base payload source implementation and interface which have to be
implemented by every specific source to be able to use them.

Set up and tear down methods are private and will not be part of the
public API. Having these as private methods will avoid race conditions
and there is basically no use for these methods outside of payload
handler.