Don't add more devices in boot-device NVRAM than the maximum allowed
The boot-device NVRAM variable has a maximum number of devices allowed,
that is specified in /sys/firmware/devicetree/base/ibm,max-boot-devices
OF property.

Adding more devices than the maximum number allowed in the boot-device
NVRAM variable can cause issues so the device list need to be restricted
to this maximum before updating the variable with nvram --update-config.

Resolves: rhbz#1748756

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>