Check source readiness in Payload install task
Currently source is implemented only for Live OS but the task is shared
between Live OS and Live Image handlers so make the source optional for
now.

In future commits usage of the source will be extended that the source
will provide also a source installation path where the installation data
can be found.