Extend the support for DBus specification
Use the class DBusSpecification to work with a representation of a
DBus specification. Use the class DBusSpecificationParser to parse
a XML specification.

The class DBusSpecificationGenerator can be used to generate a XML
specification from a class. The generated XML can be accessed with
the attribute __dbus_xml__.

The classes have to be always instantiated.