Remove payload source ready setter
Ready will be switched from a property to a dynamic checker.

This solution does not work now. The reason is that when the setup and
teardown tasks are not published then the signals if they
succeded or failed does not work. These signals are emitted only when
task is run by the start() method and not the run() method. However, the
run() method is used by payload SetUpSourcesTask.