Create variants for the signal PropertiesChanged
Transform the values of changed properties into variants by default.
It is a temporary workaround, because the variants should be provided
by the emitted signal.

The method _find_emitter has to create and return new callable objects
for the handler's methods, because the handler doesn't support weak
references that are used in our implementation of the signal.