Remove disused support for automatically fetching packages from Koji
In the past, Anaconda dependencies used to change much more often
and it was useful to automatically fetch corresponding package builds
from Koji as needed.

The Anaconda dependencies no longer change so rapidly & it is often
possible to simply use a git checkout instead for projects like
Blivet or Pykickstart.

Also, due to changes in Koji bindings, this likely no longer
works anyway.

So drop this code, it will still be in the Git history in case
we decide to review this in the future.