Return back AutoPart kickstart command (#1761901)
The AutoPart command was completely removed by the commit
d28a76fe1b19d4d0ac98b4f89020a99117278310. The command is tested in
pykickstart for collisions with other storage commands by
handler.autopart.seen call.
That works until this command was replaced by UselessCommand which
removed the handler.autopart command. It does not exists because
this handler mapping is mapped based on the __class__.__name__ so
instead of having the handler.autopart we end up with
handler.uselesscommand.

This is a hotfix for the issue to unblock installations. The correct fix
should go to pykickstart and test if handler contains autopart attribute
before touching it.

Resolves: rhbz#1761901