Fix Live Image module required value default
Having 1G as the required amount when class is created is not enough.
The reason is that when required value is not known it will return 0 as
result, in that case the old solution will set 0 as the required for
payload. However, originally this should default to 1G when size is not
known.

Fixing by implementing reasonable default from the base class.