Forbid trailing period (dot) in hostname (#1648107)
Since some time we use systemd for setting hostname immediately, and it does
not accept the dot. Systemd does this to follow kernel and only propagates
the limitation closer to user. Survey of other hostname change methods shows
this is consistently obeyed, by various means such as input sanitization or
using "pretty" hostname. For more details see the actual referenced bug.

A complication is that code previously freely mixed two kinds of host names:
These of the system to install ("local" or "static") and these of other
systems ("on the network"). The previous validation was true to DNS rules
(network) and checked if the string is a fully-qualified domain name (FQDN).
Given the restrictions in kernel+systemd, the check for local host name
additionally forbids dot.

This is solved by keeping all current names in place and adding a new
parameter to the validation function pyanaconda.network.is_valid_hostname().
All inputs for the system's (local, static) host name pass with this.

Hostname ks module validation test is also changed to match this.

Resolves: rhbz#1648107