Change flags.cmdline to use kernel.kernel_arguments
This makes sure there is only one globally shared KernelArguments instance and
if something breaks it, it will be broken equally and identically everywhere.
Note that it is possible to call any of the instance's read methods and thus
change its data, just as it already was before. This only ensures there aren't
two different potentially usable instances lying around that could be broken
independently.