keyboard: fix return value of LocaledWrapper.layouts_variants property
Localed service is working with string (comma-separated values), localization
module with list of values.

The wrapper property layouts_variants used to return [""] for "" returned by
localed service (which was ok when using the property internally in the
wrapper) while in this case the value usable by the keyboard module property
(eg returned by convert function of the wrapper) should be [].

This code will go away when we move to wrapping localed service by dasbus.