Reduce progress spinner repaints to save CPU (rhbz#1204242)
This makes the spinner redraw less (at ~24 FPS), which in turn frees CPU for
actual install work. Consequently, this speeds up testing in single-processor
virtual machines.

The savings are proportional to the time the spinner would be visible, so
the more packages, the more time saved. Thus the effect is more significant
when installing Workstation with rich software selection.

Resolves: rhbz#1204242