Add and use a mechanism for aborting early with invalid locale
This adds a function to detect if a locale or language spec is valid, and
raises an exception. The function is used in all applicable functions in
localization, eradicating many uses of parse_langcode and LANGCODE_RE.

Remove also now invalid mentions of LANGCODE_RE in method docstrings.

Apply PEP257 formatting to any touched docstrings.