Fix cppcheck problem with undefined macros
Macro G_DEFINE_TYPE and HAVE_WORKING_FORK are defined outside of our
code base and cppcheck have problem to recognize them. So we will define
this macro for the cppcheck manually.

Also add --force to force cppcheck to look even to branches which won't
be compiled with the macro defined.