Fix subscription status not being displayed in some cases
Turns out the JSON data returned by AutoAttach() is not always
available (see bug 1790931) and we can't depend on it to provide
us with information about attached subscriptions.

So let's use the GetPools() DBus method instead & call it after
each successful subscription attempt to query what subscriptions
have been attached to the system.

The JSON data returned by GetPools() is slightly different,
so the code parsing it will have to be adapted as well.

Also, when we are at it, let's also move the code querying
final System Purpose data out of the auto attach task and
move it to the new method calling GetPools(). This should
make the code more consistent.

Resolves: rhbz#1788036