Log exceptions raised in subscription data fetching method
Exceptions raised in the subscription data fetching method are
for some reason not automatically logged, so wrap the method call
explicitly with an try-except block.

This way we will be able to find out what went wrong instead of
the code simply not being executed without any feedback.

Related: rhbz#1788036