Add pylint censorship - replacement of pocketlint
After benchmark of pylint/pocketlint and investigation we have found out
that pocketlint is slow and really not useful anymore. The only useful
thing it does is filtering false positives based on regular expression.

So I've wrote a small script which does only this and behaves in a
similar way as pocketlint. Other solution would be to re-write
pocketlint but it's used by plenty of projects so it would be harder.

This script is just a temporary solution. Ideal state would be to add
this feature directly to pylint.