Fix DBus error detection
Use hasattr() instead of getattr() when trying to detect if an exception
comes from DBus. Otherwise the attempt would throw a traceback if an
unknown non-DBus exception is encountered.

Also raise exceptions that happen in tasks instead of triggering failure
limbo, so that the exception is nicely presented to the user.

Related: rhbz#1794767