Disable package download to / /tmp (#1781517)
The / and /tmp mount points aren't good candidates for package download.

During stage2 installation they are on TMPFS or squashfs. Both these
systems gives you an abstract number when you asking for a free space.
It's getting even more complicated in case when zram/zswap is involved.

During the image install you have /tmp as TMPFS and / as your root. This
is pretty similar situation, we don't want to download packages there
neither.

We are not able reliably tell user that the packages will be downloaded
correctly.

Resolves: rhbz#1781517

Reported-by: Lukas Doktor <ldoktor@redhat.com>