Fix PR tests with mock version 2
There is a new mock version which also dragging new and incompatible
configuration files. Unfortunately, this mock has a bug that you can't
install packages with the same name as local files/directories.

Add temporary workaround to fix our failing tests. This workaround will
be removed as soon as this mock issue will be resolved in the Fedora.