Remove automatic push of pot files during release
We were pushing pot files to Zanata automatically but right now it's
better to be cautions before pushing to our translation repository.
First because it's a new approach second there will be visible history
so we don't want to mess with that.

The pot file commit is automatically created but it won't be pushed. Admin is
responsible for manual check and push.

Unfortunately this manual step make release in a mock too complicated,
so remove the tutorial from release.rst document for now.