Remove auto-discover of git branch
This was used only for localization but it would be better to specify
this manually then use the script. Problem of the script is that it
will not work in 100% cases because of how git works.

TODO: We should separate these distribution specific values to one file,
which will be loaded from multiple places. This will make easier
branching and also prevent conflicts to happen between branches.