Verify that LVM destruction is orderly
This ensures that applying the `ignoredisk` Kickstart command results in a
workable combination of actions on LVM storage devices, and notifies the user
if not.

In particular, this prevents crashes in the situation where a LVM volume group
(VG) is backed by a number of physical volumes (PVs), and only some are
ignored. Destroying the VG only partially is not possible cleanly, so Anaconda
instead ends. Previously it was with a loud bang, now with at least some
semblance of dignity.

Resolves: rhbz#1688478
Resolves: rhbz#1584729
Resolves: rhbz#1698570