Avoid downloading payload image in set up phase
Before this change Anaconda will download stage2 image in the set up
phase. Result was that with big image Payload reset took ages and whole
image was downloaded to RAM without any check if the image could be
downloaded.

In my case the above resulted in this process: Anaconda started, waited
on thread and then OOM killer killed it to free up RAM. Then systemd again
started Anaconda and the whole cycle repeated in an never ending cycle.