subscription: run Subscription module on RHEL
Make sure the Subscription module is run once it's upstream code lands in RHEL in the future.

The base config file has the same list of modules already, minus the Subscription module.

As we can't append to values when we override them in the config, we need to set the whole list
here even if we want to add just a single value.