Show the help again (#1812896)
The help content is stored in directories called for example en-US, but the
valid langcode is en_US. When we try to find the best match for the current
locale in a list of langcodes, we have to provide valid langcodes. Otherwise,
no match is found and no help is provided.

Resolves: rhbz#1812896