subscription: SubscriptionRequest DBus structure
Add a SubscriptionRequest DBus structure, holding data needed
for a successful subscription attempt.

SubscriptionRequest can hold the following credentials:
- username & password
- org id & activation key/-s

As possibly both set of credentials could be set,
SubscriptionRequest has a type, that specifies which
credentials should be used for the subscription attempt.

SubscriptionRequest also holds the proxy configuration
and custom URL configuration (if any).

Red Hat account password, activation keys and HTTP proxy
password are all considered sensitive data, due to which
they are set-only. Once they are set to the Subscription
module, they will be stored internally and wiped in
SubscriptionRequest instances returned by the module.

This is acomplished by using the SecretData and
SecretData list nested DBus structures.

Using this data structure it is possible to easily wipe data
on output while still indicating it has been set.
They alsmo make it possible to wipe already set data if
requested by the client.

Also included is the InsightsEnabled property that tracks
the intent to connect the target system to Red Hat Insights.

And of course, all this is covered by extensive unit tests.