Improve help of dd_extract tool
Specify that you have to specify an absolute path to RPM file. It should
be fixed to take also relative paths but this tool is used only in the
Dracut in our scripts so this should be enough for debugging in the future.

Add FIXME because it looks that there is required parameter which is not
used at all.