Set up advanced storage on the right device tree (#1812561)
When a user enters the spoke for the advanced storage, reset the scheduled
partitioning if any to make sure that we will work with the current system’s
storage configuration.

One day, we might want to change the DBus modules and UI to work with the right
device tree, but it won't be so trivial to implement.

Resolves: rhbz#1812561