Add support for secrets in DBus structures
Use the DBus structure SecretData to represent a secret string. Use the DBus
structure SecretDataList to represent a secret list of strings. Call the
function hide_secrets to hide all secret fields of a data object. Call
the function get_public_copy to get a copy of a data object with hidden
secret fields.