Fix the encryption checkbox in the custom spoke (#1819360)
Show the checkbox for device encryption in an inconsistent state if the
device's container is encrypted. It will not trigger the on_encrypt_toggled
callback. Show the LUKS version if the checkbox is active or inconsistent.

Don't use the checkbox's method get_active for that, otherwise the callback
on_encrypt_toggled will to set the device encryption to an invalid value and
it will result in an unexpected partitioning.

Related: rhbz#1819360