Use default priority in the GUI spokes
Use old-new Standalone priority for GUI spokes so remove the priority
class attribute where it is not necessary to have it.

NetworkStandaloneSpoke had 10 because it has to be after Welcome spoke
which has priority 0 to be always first. Remove NetworkStandaloneSpoke
priority to use the default which is 100 to get the same order.