Fix TUI Kernel and Unsupported HW spokes ordering
Unsupported Hardware and Kernel Warning have the same priority so it
was based on sorted. The python 3 sorted() function states this in
documentation:

```
The built-in sorted() function is guaranteed to be stable. A sort is
stable if it guarantees not to change the relative order of elements
that compare equal — this is helpful for sorting in multiple passes (for
example, sort by department, then by salary grade).
```

Which means that the order of the input is different.

However, fixing ordering of the input field will change the required
order so instead set the priority for the Unsupported Hardware spoke.