subscription: Add SystemPurposeConfigurationTask
Add the SystemPurposeConfigurationTask that handles system purpose
configuration.

There are two cases where we need to be able to set system purpose:
- to the installation environment
- to the target system

The first one is needed so that once a subscription is attempted,
this information is used by RHSM to attach the appropriate
subscription to the system. In this case the syspurpose utility is
run in the installation environment and the system purpose data will
then be transferred by the subscription token transfer task
(to be added by future PR).

The second case happens with a user just wanting to set system purpose
for the target system, without registering and subscribing during the
installation. In this case syspurpose is not run in the installation
environment at all, but later together with the other installation
tasks, provided system purpose was not set before in the installation
environment.

The IsSystemPurposeApplied property acts a hint to avoid needlessly
setting system purpose to the target system again & possibly clobbering
system purpose adjustments RHSM might do when different system purpose data
is attached to the activation key that has been used for the subscription
attempt.