subscription: RHSM runtime configuration support
Add support for setting RHSM configuration over it's DBus
interface.

As RHSM does not provide a mechanism to restore configuration keys
to default values after they have been overwritten, we also need to
store the initial state of the RHSM config before we first start
writing into it. This way we can then restore default values if
the user changes their mind and wants Anaconda to return to a
default value after first setting a custom value for one of the
keys via the Anaconda GUI or kickstart.

In most cases this would be relevant for the server.hostname
and rhsm.baseurl keys, but it is more robust to do this for all
keys we interact with.

On the implementation side, the get_rhsm_config_defaults() method
is lazy evaluated and will cache the RHSM config state when first
called. Then SetRHMConfigurationTask can be used to set RHSM config
values & uses content of the get_rhsm_config_defaults() method in case
it needs to restore some of the values to default.