Stop passing rd.{dm,md,lvm,luks}=0 in installer environment
This blob has been maturing here ever since 2012. @dwlehman was
surprised it was still there, and thinks we've really been
aiming for anaconda/blivet to work with boot-time activation of
storage devices since 2014 or so, so this blob should go away
and we should fix any bugs that show up.

Signed-off-by: Adam Williamson <awilliam@redhat.com>