subscription: Add RestoreRHSMLogLevelTask
Used to restore the RHSM log level back to INFO. During the installation
we set the log level to DEBUG, as otherwise it is very hard to
impossible to debug subscription issues.

But this is automatically reflected in rhsm.conf, that is copied to
the target system. Due to this RHSM would log in debug mode also on
the target system.

So set the log level back to INFO, so that when the subscription token
transfer task runs rhsm.conf has log level set to INFO.