subscription: Add registration tasks
Add the username+password and organization+activation key(s) registration
tasks. Also add helper methods needed to active a private DBus session
that is needed for safe credential passing to RHSM.

For success/error reporting the tasks return to value tuple with the
first value indicating success/failure via a boolean and the second
value contains the error message, which is an empty string for success.