subscription: Add subscription related package requirements
Add package requirements for the System Purpose and Red Hat Insights
packages.

The System Purpose package is only needed if system purpose is not set
during the installation yet was requested by the user. This happens
only when no registration is attempted at installation time yet the
user still wants to set system purpose for the target system.

The Insights client package is only needed when the target system should
be connected to Red Hat Insights.