subscription: Attach subscription task
The AttachSubscription task should be run after the system
has been successfully registered. It makes sure that appropriate
subscriptions are attached to the system based on system architecture,
account used for registration and system purpose data.

Note that this effectively subscribes the running installation
environment, so that it can communicate with the CDN and install
packages from it. Separate TransferSubscriptionTokens install time
task then makes sure the subscription tokens and configuration
files are transferred to the target system.