subscription: Add check_system_purpose_set() helper function
Turns out we can easily check system purpose is set on a system by
checking if a single file exists. So create a simple helper function
that can be used from DBus tasks, the Subscription module and from
the UI. This function can then be used to replace the current property
in the Subscription module that is currently used for tracking if
System Purpose has been applied.