subscription: Add AttachedSubscription DBus structure
The AttachedSubscription DBus structure describes a single subscription
that has been attached to the target system.

A list of AttachedSubscription instances together with the InsightsEnabled property
and SystemPurposeData property describe the system state post registration and subscription.

Note that all the fields are effectively string "blobs" we are getting
from a JSON file returned by RHSM DBus API.

Only for the consumed_entitlement_count field we can be reasonably sure there will
always be something in the JSON we can parse into an integer.