subscription: Add glade file for Subscription spoke
This is the full Glade file for Subscription spoke taken
from the RHEL 8.2 Anaconda codebase. The only change is
a couple lines adding change signal to widgets that had
none in 8.2.

Given that Glade is pretty much compiled XML assembly,
there is not really any option to move this piece by piece.

On the other hand we don't expect any layout changes as part
of moving the Subscription spoke upstream, so adding the known
working Glade file like this should be good enough.