subscription: Add main Subscription spoke structure
This is mostly boilerplate needed to communicate with and manipulate
all the various entry fields, check boxes, buttons and other controls
present on the Subscription spoke.

There is also the assortment of common spoke methods and properties.
The biggest of those is the initialize method that grabs reference
to widgets from the parsed glade file and does initial configuration of
the spoke.

Pieces of the spoke have been left out as they depend on asynchronous
registration/unregistration support (WiP), source switching (WiP) or
simply to not make the pull request even larger.

A short list of these parts that have been left out and are expected
to be added by followup (much smaller) pull requests:
- status message generation
- network connectivity check
- handler for the register/unregister button
- code for interacting with the asynchronous register/unregister thread
- code for displaying attached subscriptions
- interaction with source switching to make CDN the default installation
  or to switch back to best non CDN installation source