subscription: Simplify system purpose configuration
Turn out is it fully sufficient to set the system purpose data in the
installation environment. Later at installation time the
TransferSubscriptionTokensTask will transfer the syspurpose.json file
from the installation environment to the target system, if one exists.

It is still needed to run the SystemPurposeConfigurationTask task
to apply the data, but it still makes it possible to simplify the
code quite a bit:
- we don't need to run SystemPurposeConfigurationTask at install time at
  all, which means we also never need to add the syspurpose tool package
  to the installation transaction
- we can drop the IsSystemPurposeApplied property as it is now needed
  in a minimum of places and a simple check for the syspurpose.json file
  existing in the given sysroot is more than enough
- we can more SystemPurposeConfigurationTask to runtime.py
  as it is now effectively a runtime only task